requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
langchain
langchain-community
langchain_openai
langchain-mistralai
langchain-google-genai
langchain-pinecone
fireworks-ai
langchain-groq
gigachat
pillow
httpx[socks]
redis
celery[redis]
aioredis
elasticsearch